Refactor to use `timeupdate` rather than `cuechange`
There doesn’t appear to be a way to reliably detect when the
user disables subtitles using the stock player controls and
on all browsers tested the controls will set the track mode
to `disabled` rather than what it was before subtitle
display was enabled (`hidden`).

Instead, we’ll simply use the video element’s `timeupdate`
events which fire more frequently but aren’t tied to the
track mode.